New algorithm for findRoots
Dont actually factor out the roots in the search polynom p but add them
to a second polynom q and search roots on the quotient p(z) / q(z).
For N=50 the old algorithm did not finish to find all roots of the
nominator polynom p.
After finding a root, it is added to q, so for the next search, the
actual position of z has to be updated, or a division by zero occures.
It was found, that a shift in the imaginary direction works well, while
a shift in the real direction could lead to oscillations.